requirements.txt
streamlit
openai
pillow
pymupdf
python-docx
pandas
xlsxwriter
